chore(benches): remove redundant hnsw_benchmark.rs (#425)
- Move bench_hnsw_construction into vector_search_bench.rs alongside the
  existing Flat/IVF construction benches, applying the #427 hygiene rules
  (deterministic seeds, SAMPLE_SIZE_SLOW, Throughput::Elements).
- Delete benches/hnsw_benchmark.rs in full; HNSW search was already covered
  by vector_search_bench.rs::bench_hnsw_search_compare.
- Drop the [[bench]] name = "hnsw_benchmark" entry from laurus/Cargo.toml.

Net change: -72 lines, single source of truth for vector index benches.

Closes #425
